+++ src/parens.cpp
@@ -59,7 +59,7 @@ void do_parens(void)
          }
 
          /* Grab the close sparen */
-         chunk_t *pclose = chunk_get_next_type(pc, CT_SPAREN_CLOSE, pc->level, CNAV_PREPROC);
+         chunk_t *pclose = chunk_get_next_type(pc, CT_SPAREN_CLOSE, (int)pc->level, CNAV_PREPROC);
          if (pclose != NULL)
          {
             check_bool_parens(pc, pclose, 0);
